+++ src/Ext/RadSite/Body.h
@@ -42,6 +42,8 @@ public:
 
 		virtual void LoadFromStream(PhobosStreamReader& Stm) override;
 		virtual void SaveToStream(PhobosStreamWriter& Stm) override;
+		virtual void Initialize() override;
+
 	private:
 		template <typename T>
 		void Serialize(T& Stm);