+++ src/Ext/RadSite/Hooks.cpp
@@ -11,6 +11,7 @@
 #include <Ext/Rules/Body.h>
 #include <Ext/Techno/Body.h>
 
+#include <Utilities/Macro.h>
 /*
 	Custom Radiations
 	Worked out from old uncommented Ares RadSite Hook , adding some more hook
@@ -35,7 +36,7 @@ DEFINE_HOOK(0x469150, BulletClass_Detonate_ApplyRadiation, 0x5)
 		auto const pExt = BulletExt::ExtMap.Find(pThis);
 		auto const pWH = pThis->WH;
 		auto const cell = CellClass::Coord2Cell(*pCoords);
-		auto const spread = static_cast<int>(pWH->CellSpread);
+		auto const spread = Game::F2I(pWH->CellSpread);
 
 		pExt->ApplyRadiationToCell(cell, spread, pWeapon->RadLevel);
 	}
@@ -50,7 +51,7 @@ DEFINE_HOOK(0x46ADE0, BulletClass_ApplyRadiation_UnUsed, 0x5)
 	return 0x46AE5E;
 }
 
-// Fix for desolator unable to fire his deploy weapon when cloaked 
+// Fix for desolator 
 DEFINE_HOOK(0x5213E3, InfantryClass_AIDeployment_CheckRad, 0x4)
 {
 	GET(InfantryClass*, pInf, ESI);
@@ -68,13 +69,13 @@ DEFINE_HOOK(0x5213E3, InfantryClass_AIDeployment_CheckRad, 0x4)
 
 		auto const it = std::find_if(RadSiteExt::Array.begin(), RadSiteExt::Array.end(),
 			[=](RadSiteExt::ExtData* const pSite)
-		{
-			return
-				pSite->Type == pRadType &&
-				pSite->OwnerObject()->BaseCell == currentCoord &&
-				pSite->OwnerObject()->Spread == static_cast<int>(warhead->CellSpread)
-				;
-		});
+			{
+				return
+					pSite->Type == pRadType &&
+					pSite->OwnerObject()->BaseCell == currentCoord &&
+					pSite->OwnerObject()->Spread == Game::F2I(warhead->CellSpread)
+					;
+			});
 
 		if (it != RadSiteExt::Array.end())
 		{
@@ -88,6 +89,7 @@ DEFINE_HOOK(0x5213E3, InfantryClass_AIDeployment_CheckRad, 0x4)
 		0x5213F4 : 0x521484;
 }
 
+// Fix for desolator unable to fire his deploy weapon when cloaked 
 DEFINE_HOOK(0x521478, InfantryClass_AIDeployment_FireNotOKCloakFix, 0x4)
 {
 	GET(InfantryClass* const, pThis, ESI);
@@ -95,13 +97,11 @@ DEFINE_HOOK(0x521478, InfantryClass_AIDeployment_FireNotOKCloakFix, 0x4)
 	auto const pWeapon = pThis->GetDeployWeapon()->WeaponType;
 	AbstractClass* pTarget = nullptr; //default WWP nullptr
 
-	if (pWeapon 
+	if (pWeapon
 		&& pWeapon->DecloakToFire
-		&& pWeapon->AreaFire
-		&& pThis->Type->DeployFire
 		&& (pThis->CloakState == CloakState::Cloaked || pThis->CloakState == CloakState::Cloaking))
 	{
-		pTarget = pThis->GetCell(); //target it own cell
+		pTarget = MapClass::Instance->TryGetCellAt(pThis->GetCoords());
 	}
 
 	pThis->SetTarget(pTarget); //Here we go
@@ -114,21 +114,21 @@ DEFINE_HOOK(0x43FB23, BuildingClass_AI, 0x5)
 {
 	GET(BuildingClass* const, pBuilding, ECX);
 
-	if (pBuilding->IsIronCurtained() || pBuilding->Type->ImmuneToRadiation || pBuilding->InLimbo || pBuilding->BeingWarpedOut)
+	if (pBuilding->IsIronCurtained() || pBuilding->Type->ImmuneToRadiation || pBuilding->InLimbo || pBuilding->BeingWarpedOut || pBuilding->TemporalTargetingMe)
 		return 0;
 
 	auto const buildingCoords = pBuilding->GetMapCoords();
-	for (auto pFoundation = pBuilding->GetFoundationData(false); *pFoundation != CellStruct { 0x7FFF, 0x7FFF }; ++pFoundation)
+	for (auto pFoundation = pBuilding->GetFoundationData(false); *pFoundation != CellStruct{ 0x7FFF, 0x7FFF }; ++pFoundation)
 	{
-		CellStruct CurrentCoord = buildingCoords + *pFoundation;
+		CellStruct nCurrentCoord = buildingCoords + *pFoundation;
 
-		for (auto pRadExt : RadSiteExt::Array)
+		for (auto& pRadExt : RadSiteExt::Array)
 		{
 			RadSiteClass* pRadSite = pRadExt->OwnerObject();
 			RadTypeClass* pType = pRadExt->Type;
 
 			// Check the distance, if not in range, just skip this one
-			double orDistance = pRadSite->BaseCell.DistanceFrom(CurrentCoord);
+			double orDistance = pRadSite->BaseCell.DistanceFrom(nCurrentCoord);
 			if (pRadSite->Spread < orDistance - 0.5)
 				continue;
 
@@ -136,35 +136,37 @@ DEFINE_HOOK(0x43FB23, BuildingClass_AI, 0x5)
 			if ((delay == 0) || (Unsorted::CurrentFrame % delay != 0))
 				continue;
 
-			if (RadSiteExt::GetRadLevelAt(pRadSite,CurrentCoord) <= 0.0 || !pType->GetWarhead())
+			if (RadSiteExt::GetRadLevelAt(pRadSite, nCurrentCoord) <= 0.0 || !pType->GetWarhead())
 				continue;
 
-			auto warhead = pType->GetWarhead();
-			auto absolute = warhead->WallAbsoluteDestroyer;
+			auto pWarhead = pType->GetWarhead();
+			auto absolute = pWarhead->WallAbsoluteDestroyer;
 			bool ignore = pBuilding->Type->Wall && absolute;
-			auto damage = static_cast<int>((RadSiteExt::GetRadLevelAt(pRadSite, CurrentCoord) / 2) * pType->GetLevelFactor());
+			auto damage = Game::F2I((RadSiteExt::GetRadLevelAt(pRadSite, nCurrentCoord) / 2) * pType->GetLevelFactor());
 
 			if (pBuilding->IsAlive) // simple fix for previous issues
-				pBuilding->ReceiveDamage(&damage, static_cast<int>(orDistance), warhead, nullptr, ignore, absolute, pRadExt->RadHouse.Get());
+				if (pBuilding->ReceiveDamage(&damage, Game::F2I(orDistance), pWarhead, nullptr, ignore, absolute, pRadExt->RadHouse.Get()) == DamageState::NowDead)
+					break; //dont continue , meaningless
 		}
 	}
 
 	return 0;
 }
 
-// be aware that this function is updated every frame 
-// putting debug log here can become mess because it gonna print bunch of debug line
-DEFINE_HOOK(0x4DA554, FootClass_AI_RadSiteClass, 0x5)
+// skip Frame % RadApplicationDelay
+DEFINE_LJMP(0x4DA554, 0x4DA56E);
+
+// Hook Adjusted to support Ares RadImmune Ability check
+DEFINE_HOOK(0x4DA59F, FootClass_AI_Radiation, 0x5)
 {
 	GET(FootClass* const, pFoot, ESI);
-	auto state = static_cast<DamageState>(R->AL());
 
-	if (!pFoot->IsIronCurtained() && !pFoot->GetTechnoType()->ImmuneToRadiation && !pFoot->InLimbo && !pFoot->IsInAir())
+	if (!pFoot->IsIronCurtained() && pFoot->IsInPlayfield && !pFoot->TemporalTargetingMe)
 	{
 		CellStruct CurrentCoord = pFoot->GetCell()->MapCoords;
 
 		// Loop for each different radiation stored in the RadSites container
-		for (auto const pRadExt : RadSiteExt::Array)
+		for (auto& pRadExt : RadSiteExt::Array)
 		{
 			RadSiteClass* pRadSite = pRadExt->OwnerObject();
 
@@ -179,21 +181,22 @@ DEFINE_HOOK(0x4DA554, FootClass_AI_RadSiteClass, 0x5)
 				continue;
 
 			// for more precise dmg calculation
-			double RadLevel = RadSiteExt::GetRadLevelAt(pRadSite, CurrentCoord);
-			if (RadLevel <= 0.0 || !pType->GetWarhead())
+			double nRadLevel = RadSiteExt::GetRadLevelAt(pRadSite, CurrentCoord);
+			if (nRadLevel <= 0.0 || !pType->GetWarhead())
 				continue;
 
-			int damage = static_cast<int>(RadLevel * pType->GetLevelFactor());
-			int distance = static_cast<int>(orDistance);
-			auto warhead = pType->GetWarhead();
-			auto absolute = warhead->WallAbsoluteDestroyer;
+			int damage = Game::F2I(nRadLevel * pType->GetLevelFactor());
+			int distance = Game::F2I(orDistance);
+			auto pWarhead = pType->GetWarhead();
+			auto absolute = pWarhead->WallAbsoluteDestroyer;
 
-			if (pFoot->IsAlive || !pFoot->IsSinking)// simple fix for previous issues
-				state = pFoot->ReceiveDamage(&damage, distance, warhead, nullptr, false, absolute, pRadExt->RadHouse.Get());
+			if (pFoot->IsAlive || !pFoot->IsSinking)
+			{
+				if (pFoot->ReceiveDamage(&damage, distance, pWarhead, nullptr, false, absolute, pRadExt->RadHouse.Get()) == DamageState::NowDead)
+					break; //dont continue , meaningless
+			}
 		}
 	}
-
-	R->EAX<DamageState>(state);
 
 	return pFoot->IsAlive ? 0x4DA63B : 0x4DAF00;
 }
@@ -203,8 +206,8 @@ DEFINE_HOOK(0x4DA554, FootClass_AI_RadSiteClass, 0x5)
 	RadSiteExt::ExtData* pExt = RadSiteExt::ExtMap.Find(pThis);\
 	auto output = pExt->Type->## value ##;
 
-/* 
-//All part of 0x65B580 Hooks is here 
+/*
+//All part of 0x65B580 Hooks is here
 DEFINE_HOOK(65B593, RadSiteClass_Activate_Delay, 6)
 {
 	GET(RadSiteClass* const, pThis, ECX);